tests/wpt/web-platform-tests/shared-storage/permissions-policy-none.tentative.https.sub.html

<!doctype html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script src=/permissions-policy/resources/permissions-policy.js></script>
  <script src="/shared-storage/resources/util.js"></script>
  <script>
    'use strict';
    const same_origin_src = '/shared-storage/resources/permissions-policy-helper.html';
    const cross_origin_src = 'https://{{domains[www]}}:{{ports[https][0]}}' +
      same_origin_src;
    const header = 'permissions policy header shared-storage=()';

    promise_test(async t => {
      const allowed = await AreSharedStorageMethodsAllowedByPermissionsPolicy();
      assert_false(allowed);
    }, header + ' disallows sharedStorage in the current page.');

    async_test(t => {
      test_feature_availability('shared-storage', t, same_origin_src,
          expect_feature_unavailable_default);
    }, header + ' disallows sharedStorage in same-origin iframes.');

    async_test(t => {
      test_feature_availability('shared-storage', t, cross_origin_src,
          expect_feature_unavailable_default);
    }, header + ' disallows sharedStorage in cross-origin iframes.');
  </script>
</body>
